fix random downsample type bug (#1108)

+++ sensing/preprocessor/pointcloud/pointcloud_preprocessor/include/pointcloud_preprocessor/downsample_filter/random_downsample_filter_nodelet.h
@@ -73,7 +73,7 @@ protected:
     pointcloud_preprocessor::RandomDownsampleFilterConfig & config, uint32_t level);
 
 private:
-  double sample_num_;
+  size_t sample_num_;
 
 public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
